requirements.txt
# Django
Django==5.1.2
djangorestframework==3.15.2


# Photo & Images
pillow==10.4.0

# Authentication


# Database